requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic==2.9.2
uvicorn[standard]
requests
pytest
supabase
pyjwt
passlib[bcrypt]
pika
redis
celery[redis]
fastapi-jwt-auth[asymmetric]
python-dotenv